# dto/admin_login.go
package dto

import (
	"github.com/gin-gonic/gin"
	"github.com/starMoonZhao/go_gateway/public"
	"time"
)

type AdminLoginInput struct {
	UserName string `json:"username" form:"username" comment:"姓名" example:"admin" validate:"required"`
	Password string `json:"password" form:"password" comment:"密码" example:"123456" validate:"required"`
}

// 校验用户输入的登陆值
func (param *AdminLoginInput) BindValidParam(c *gin.Context) error {
	return public.DefaultGetValidParams(c, param)
}

type AdminLoginOutput struct {
	Token string `json:"token" form:"token" comment:"token" example:"token" validate:""`
}

type AdminSessionInfo struct {
	Id        int64     `json:"id"`
	UserName  string    `json:"username"`
	LoginTime time.Time `json:"login_time"`
}